It's been observed that in the previous code, if some exception occurs after the zip files has been committed and streamed to the client, the thrown exception could disrupt the output streaming and causing the zip file to be corrupt.  Although the request validity checkers try to actively detect such problems prior to committing the output stream to the client, there is however still a slim chance this could occur, e.g. while streaming data to the client, the cassandra server crashes or become unreachable.
So the only solution is to do graceful degradation that if something happens, the data api should make sure the zip file is properly closed without corruption.  as an added feature, the code also attempts to inject a special entry into the zip file, ERROR.err, to notify the client of the error, and warn the client that the received zip file can be incomplete and missing entries.

All ZipMaker implementations now catch exceptions to gracefully close the zip stream to the client, and also injects the ERROR.err entry.

Added a static nested Helper class to the ZipMakerFactory with a static method for injecting ERROR.err into the zip outputstream